+++ src/lib.rs
@@ -25,7 +25,7 @@ use std::ptr::NonNull;
 
 const PACKB_DOC: &str =
     "packb(obj, /, default=None, option=None)\n--\n\nSerialize Python objects to msgpack.\0";
-const UNPACKB_DOC: &str = "unpackb(obj, /)\n--\n\nDeserialize msgpack to Python objects.\0";
+const UNPACKB_DOC: &str = "unpackb(obj, /, option=None)\n--\n\nDeserialize msgpack to Python objects.\0";
 
 macro_rules! opt {
     ($mptr:expr, $name:expr, $opt:expr) => {
@@ -56,6 +56,7 @@ pub unsafe extern "C" fn PyInit_ormsgpack() -> *mut PyObject {
     };
 
     let wrapped_packb: PyMethodDef;
+    let wrapped_unpackb: PyMethodDef;
 
     #[cfg(python37)]
     {
@@ -93,12 +94,30 @@ pub unsafe extern "C" fn PyInit_ormsgpack() -> *mut PyObject {
         )
     };
 
-    let wrapped_unpackb = PyMethodDef {
-        ml_name: "unpackb\0".as_ptr() as *const c_char,
-        ml_meth: Some(unpackb),
-        ml_flags: METH_O,
-        ml_doc: UNPACKB_DOC.as_ptr() as *const c_char,
-    };
+
+    #[cfg(python37)]
+    {
+        wrapped_unpackb = PyMethodDef {
+            ml_name: "unpackb\0".as_ptr() as *const c_char,
+            ml_meth: Some(unsafe {
+                std::mem::transmute::<pyo3::ffi::_PyCFunctionFastWithKeywords, PyCFunction>(unpackb)
+            }),
+            ml_flags: pyo3::ffi::METH_FASTCALL | METH_KEYWORDS,
+            ml_doc: UNPACKB_DOC.as_ptr() as *const c_char,
+        };
+    }
+
+    #[cfg(not(python37))]
+    {
+        wrapped_unpackb = PyMethodDef {
+            ml_name: "unpackb\0".as_ptr() as *const c_char,
+            ml_meth: Some(unsafe {
+                std::mem::transmute::<PyCFunctionWithKeywords, PyCFunction>(unpackb)
+            }),
+            ml_flags: METH_VARARGS | METH_KEYWORDS,
+            ml_doc: UNPACKB_DOC.as_ptr() as *const c_char,
+        };
+    }
 
     unsafe {
         PyModule_AddObject(
@@ -179,13 +198,126 @@ fn raise_packb_exception(msg: Cow<str>) -> *mut PyObject {
     std::ptr::null_mut()
 }
 
+
+#[cfg(python37)]
 #[no_mangle]
-pub unsafe extern "C" fn unpackb(_self: *mut PyObject, obj: *mut PyObject) -> *mut PyObject {
-    match crate::deserialize::deserialize(obj) {
+pub unsafe extern "C" fn unpackb(
+    _self: *mut PyObject,
+    args: *const *mut PyObject,
+    nargs: Py_ssize_t,
+    kwnames: *mut PyObject,
+) -> *mut PyObject {
+    let mut optsptr: Option<NonNull<PyObject>> = None;
+
+    let num_args = pyo3::ffi::PyVectorcall_NARGS(nargs as usize);
+    if unlikely!(num_args != 1) {
+        let msg;
+        if num_args > 1 {
+            msg = Cow::Borrowed(
+                "unpackb() accepts only 1 positional argument",
+            );
+        }
+        else {
+            msg = Cow::Borrowed(
+                "unpackb() missing 1 required positional argument: 'obj'",
+            )
+        }
+        return raise_unpackb_exception(deserialize::DeserializeError::new(msg));
+    }
+    if !kwnames.is_null() {
+        for i in 0..=PyTuple_GET_SIZE(kwnames) - 1 {
+            let arg = PyTuple_GET_ITEM(kwnames, i as Py_ssize_t);
+            if arg == typeref::OPTION {
+                optsptr = Some(NonNull::new_unchecked(*args.offset(num_args + i)));
+            } else {
+                return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed(
+                    "unpackb() got an unexpected keyword argument",
+                )));
+            }
+        }
+    }
+
+    let mut optsbits: i32 = 0;
+    if let Some(opts) = optsptr {
+        if (*opts.as_ptr()).ob_type != typeref::INT_TYPE {
+            return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed("Invalid opts")));
+        }
+        optsbits = PyLong_AsLong(optsptr.unwrap().as_ptr()) as i32;
+        if !(0..=opt::MAX_UNPACKB_OPT).contains(&optsbits) {
+            return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed("Invalid opts")));
+        }
+    }
+
+    match crate::deserialize::deserialize(*args, optsbits as opt::Opt) {
         Ok(val) => val.as_ptr(),
         Err(err) => raise_unpackb_exception(err),
     }
 }
+
+#[cfg(not(python37))]
+#[no_mangle]
+pub unsafe extern "C" fn unpackb(
+    _self: *mut PyObject,
+    args: *mut PyObject,
+    kwds: *mut PyObject,
+) -> *mut PyObject {
+    let mut optsptr: Option<NonNull<PyObject>> = None;
+
+    let obj = PyTuple_GET_ITEM(args, 0);
+
+    let num_args = PyTuple_GET_SIZE(args);
+    if unlikely!(num_args != 1) {
+        let msg;
+        if num_args > 1 {
+            msg = Cow::Borrowed(
+                "unpackb() accepts only 1 positional argument",
+            );
+        }
+        else {
+            msg = Cow::Borrowed(
+                "unpackb() missing 1 required positional argument: 'obj'",
+            )
+        }
+        return raise_unpackb_exception(msg);
+    }
+
+    if !kwds.is_null() {
+        let len = unsafe { crate::ffi::PyDict_GET_SIZE(kwds) as usize };
+        let mut pos = 0isize;
+        let mut arg: *mut PyObject = std::ptr::null_mut();
+        let mut val: *mut PyObject = std::ptr::null_mut();
+        for _ in 0..=len - 1 {
+            unsafe { _PyDict_Next(kwds, &mut pos, &mut arg, &mut val, std::ptr::null_mut()) };
+            if arg == typeref::OPTION {
+                optsptr = Some(NonNull::new_unchecked(val));
+            } else if arg.is_null() {
+                break;
+            } else {
+                return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed(
+                    "unpackb() got an unexpected keyword argument")
+                ));
+            }
+        }
+    }
+
+    let mut optsbits: i32 = 0;
+    if let Some(opts) = optsptr {
+        if (*opts.as_ptr()).ob_type != typeref::INT_TYPE {
+            return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed("Invalid opts")));
+        }
+        optsbits = PyLong_AsLong(optsptr.unwrap().as_ptr()) as i32;
+        if optsbits < 0 || optsbits > opt::MAX_UNPACKB_OPT {
+            return raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Borrowed("Invalid opts")));
+        }
+    }
+
+    match crate::deserialize::deserialize(obj,optsbits as opt::Opt) {
+        Ok(val) => val.as_ptr(),
+        Err(err) => raise_unpackb_exception(deserialize::DeserializeError::new(Cow::Owned(err))),
+    }
+}
+
+
 
 #[cfg(python37)]
 #[no_mangle]
@@ -241,7 +373,7 @@ pub unsafe extern "C" fn packb(
             return raise_packb_exception(Cow::Borrowed("Invalid opts"));
         }
         optsbits = PyLong_AsLong(optsptr.unwrap().as_ptr()) as i32;
-        if !(0..=opt::MAX_OPT).contains(&optsbits) {
+        if !(0..=opt::MAX_PACKB_OPT).contains(&optsbits) {
             return raise_packb_exception(Cow::Borrowed("Invalid opts"));
         }
     }
@@ -314,7 +446,7 @@ pub unsafe extern "C" fn packb(
             return raise_packb_exception(Cow::Borrowed("Invalid opts"));
         }
         optsbits = PyLong_AsLong(optsptr.unwrap().as_ptr()) as i32;
-        if optsbits < 0 || optsbits > opt::MAX_OPT {
+        if optsbits < 0 || optsbits > opt::MAX_PACKB_OPT {
             return raise_packb_exception(Cow::Borrowed("Invalid opts"));
         }
     }