+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 ## Next Version
+### Added
+- `unpackb` now accepts keyword argument `option` with argument `OPT_NON_STR_KEYS`. This option will let ormsgpack
+    unpack dictionaries with non-str keys.
+    Be aware that this option is considered unsafe and disabled by default in msgpack due to possibility of HashDoS.
 ### Misc
 - Grouped benchmarks in a pattern that should make more sense.
 - Added pydantic docs to `README.md`